requirements.txt
requests
python-docx
tkinterweb
markdown
